requirements.txt
streamlit
mysql-connector-python
langchain-core
langchain-community
python-dotenv
pandas
google-ai-generativelanguage
google-generativeai
openpyxl
xlrd
numpy
uuid